Rename to app-integration, split utils
