+++ examples/hosts/app-integration/.editorconfig
