+++ examples/hosts/app-integration/.gitignore
