+++ examples/hosts/app-integration/.npmignore
