+++ examples/hosts/app-integration/README.md
