+++ examples/hosts/app-integration/package.json
