+++ examples/hosts/app-integration/src/app/app.ts
@@ -7,7 +7,7 @@ import { IComponent } from "@fluidframework/component-core-interfaces";
 import { Container } from "@fluidframework/container-loader";
 import { HTMLViewAdapter } from "@fluidframework/view-adapters";
 import { IComponentMountableView } from "@fluidframework/view-interfaces";
-import { getTinyliciousContainer } from "./tinyliciousContainer";
+import { getTinyliciousContainer } from "../util";
 
 // I'm choosing to put the docId in the hash just for my own convenience.  There should be no requirements on the
 // page's URL format deeper in the system.
