+++ examples/hosts/app-integration/src/app/index.html
