+++ examples/hosts/app-integration/src/component/index.ts
