+++ examples/hosts/app-integration/src/component/interface.ts
