+++ examples/hosts/app-integration/src/component/model.ts
