+++ examples/hosts/app-integration/src/component/view.tsx
