+++ examples/hosts/app-integration/src/container/index.ts
