+++ examples/hosts/app-integration/src/index.ts
