+++ examples/hosts/app-integration/src/util/getTinyliciousContainer.ts
