+++ examples/hosts/app-integration/tsconfig.esnext.json
