+++ examples/hosts/app-integration/tsconfig.json
