+++ examples/hosts/app-integration/webpack.config.js
