+++ examples/hosts/app-integration/webpack.dev.js
