+++ examples/hosts/app-integration/webpack.prod.js
